Significant speedup for YUV/NV12/P010 and related codecs
Log: replace old C-style format with modern C++20 for compile-time safety

+++ sources/effects/Animation4Music_WavesPulseFast.cpp
@@ -94,7 +94,7 @@ bool Animation4Music_WavesPulseFast::getImage(Image<ColorRgb>& newImage)
 		}
 		else
 		{
-			w2 = std::min((int)((y * (_buffer.size() - 1)) % (scaleY / 2)) * 100 / (scaleY / 2), (int)100);
+			w2 = std::min((int)((y * (_buffer.size() - 1)) % (scaleY / 2)) * 100 / (scaleY / 2), 100);
 			w1 = 100 - w2;
 		}
 		selected = ColorRgb(
@@ -127,7 +127,7 @@ bool Animation4Music_WavesPulseFast::getImage(Image<ColorRgb>& newImage)
 		}
 		else
 		{
-			we2 = std::min((int)((x * (_buffer.size() - 1)) % (scaleX / 2)) * 100 / (scaleX / 2), (int)100);
+			we2 = std::min((int)((x * (_buffer.size() - 1)) % (scaleX / 2)) * 100 / (scaleX / 2), 100);
 			we1 = 100 - we2;
 		}
 		selected = ColorRgb(
